feat(131553): Adiciona refatoração nas melhorias da agenda.
